moved checkEvSel funcion to dedicated utility header file (to be expanded in future PRs)
